refactor: move FoldingRangeAdapter to new file
