+++ src/language/common/lspLanguageFeatures.ts
@@ -39,59 +39,7 @@ export * from './DocumentFormattingEditProvider';
 
 export * from './DocumentColorAdapter';
 
-//#region FoldingRangeAdapter
-
-export interface ILanguageWorkerWithFoldingRanges {
-	getFoldingRanges(uri: string, context?: { rangeLimit?: number }): Promise<lsTypes.FoldingRange[]>;
-}
-
-export class FoldingRangeAdapter<T extends ILanguageWorkerWithFoldingRanges>
-	implements languages.FoldingRangeProvider
-{
-	constructor(private _worker: WorkerAccessor<T>) {}
-
-	public provideFoldingRanges(
-		model: editor.IReadOnlyModel,
-		context: languages.FoldingContext,
-		token: CancellationToken
-	): Promise<languages.FoldingRange[] | undefined> {
-		const resource = model.uri;
-
-		return this._worker(resource)
-			.then((worker) => worker.getFoldingRanges(resource.toString(), context))
-			.then((ranges) => {
-				if (!ranges) {
-					return;
-				}
-				return ranges.map((range) => {
-					const result: languages.FoldingRange = {
-						start: range.startLine + 1,
-						end: range.endLine + 1
-					};
-					if (typeof range.kind !== 'undefined') {
-						result.kind = toFoldingRangeKind(<lsTypes.FoldingRangeKind>range.kind);
-					}
-					return result;
-				});
-			});
-	}
-}
-
-function toFoldingRangeKind(
-	kind: lsTypes.FoldingRangeKind
-): languages.FoldingRangeKind | undefined {
-	switch (kind) {
-		case lsTypes.FoldingRangeKind.Comment:
-			return languages.FoldingRangeKind.Comment;
-		case lsTypes.FoldingRangeKind.Imports:
-			return languages.FoldingRangeKind.Imports;
-		case lsTypes.FoldingRangeKind.Region:
-			return languages.FoldingRangeKind.Region;
-	}
-	return void 0;
-}
-
-//#endregion
+export * from './FoldingRangeAdapter';
 
 //#region SelectionRangeAdapter
